crawler/requirements.txt
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
dnspython==2.4.2
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
outcome==1.3.0.post0
pycparser==2.21
pymongo==4.6.0
PySocks==1.7.1
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.1.0
wsproto==1.2.0
python-dotenv==1.0.1